requirements.txt
gradio>=5.12.0,<6.0
librosa
google-genai==1.20.0
numpy
boto3
python-dotenv
gradio-webrtc
fastrtc
google-api-python-client
google-auth-httplib2
google-auth-oauthlib



